requirements.txt
Django=>2.0
django-el-pagination==3.0.1
django-markdown-deux==1.0.5
django-simple-math-captcha==1.0.3
django-pagedown==1.0.4
